Approve Leave Of different Types

+++ EmployeeLeaveManagementApp/Domain/ApproveLeave.cs
@@ -16,7 +16,7 @@ namespace LMS_WebAPP_Domain
         public System.DateTime ToDate { get; set; }
         public System.DateTime CreatedDate { get; set; }
         public int RefStatus { get; set; }
-        public int NumberOfWorkingDays { get; set; }
+        public double NumberOfWorkingDays { get; set; }
         public int RefLeaveType { get; set; }
         public string EmployeeComment { get; set; }
         public string ManagerComments { get; set; }
